refactor: extract readFileBytes helper + wrap makeDsn and envelope creation in try/catch
- Centralise ENOENT/IO error handling into shared readFileBytes() in transport.ts
- Removes duplicated file-reading error block from both send-event.ts and send-envelope.ts
- Wrap makeDsn() call in try/catch to guard against future SDK internal throws
- Wrap createEventEnvelope()+serializeEnvelope() in try/catch with descriptive ValidationError

Addresses Sentry Seer findings (medium) and Cursor Bugbot finding (low) on PR #921.
